compartmentalize playpage
