+++ src/app/play/[id]/page.tsx
@@ -5,26 +5,22 @@ import React, { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
 
 // types
-import { ActivePlayer, PlayGame } from "@/types/globals";
+import { PlayGame } from "@/types/globals";
 
 // components
-import { ChevronLeftIcon, ChevronRightIcon } from "@heroicons/react/24/solid";
 import { Button } from "@/components/ui/button";
 import {
   ResizableHandle,
   ResizablePanel,
   ResizablePanelGroup,
 } from "@/components/ui/resizable";
-import { Input } from "@/components/ui/input";
 import Link from "next/link";
-import EndGame from "@/components/EndGame";
-import Bar from "@/components/Bar";
-import { Settings } from "lucide-react";
 
 // utils
 import { handleGameState } from "@/lib/functions/gamelogic";
 import { useStoreTable } from "@/store/tablesStore";
-import SettingsPlay from "@/components/SettingsPlay";
+import PlayLeftPanel from "./PlayLeftPanel";
+import PlayRightPanel from "./PlayRightPanel";
 
 export default function Play({ params }: { params: { id: string } }) {
   // find table from storage, if not redirect
@@ -34,105 +30,44 @@ export default function Play({ params }: { params: { id: string } }) {
   if (!findTable) router.replace("/404");
   const table = findTable as PlayGame;
 
-  //enable or disable auto layout of panels.
+  const [tableFinished, setTableFinished] = useState(table.tableFinished);
   const [autoLayout, setAutoLayout] = useState(true);
   const [panelDirection, setPanelDirection] = useState<
     "vertical" | "horizontal"
   >("vertical");
 
-  // table states
-  const [turns, setTurns] = useState(table.rounds);
-  const [currentTurn, setCurrentTurn] = useState(table.rounds);
-  const [tableFinished, setTableFinished] = useState(table.tableFinished);
-  const [activePlayers, setActivePlayers] = useState<ActivePlayer[]>(
-    table.players
-  );
-  const [highestScore, setHighestScore] = useState(0);
-  const [minScore, setMinScore] = useState(0);
-
+  //enable or disable auto layout of panels.
   useEffect(() => {
-    updateHighScore();
-
-    setPanelDirection(windowResizeHandler);
+    windowResizeHandler();
     window.addEventListener("resize", () => {
-      setPanelDirection(windowResizeHandler);
+      windowResizeHandler();
     });
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   function windowResizeHandler() {
-    if (!autoLayout) return panelDirection;
-    return window.innerWidth > 768 ? "horizontal" : "vertical";
+    switch (screen.orientation.type) {
+      case "landscape-primary":
+      case "landscape-secondary":
+        setPanelDirection("horizontal");
+        break;
+      case "portrait-primary":
+      case "portrait-secondary":
+        setPanelDirection("vertical");
+        break;
+    }
   }
 
-  const updateHighScore = () => {
-    let max = 0;
-    let min = 0;
+  // update page when table is finished
 
-    table.players.forEach((player) => {
-      if (player.total > max) max = player.total;
-      if (player.total < min) min = player.total;
-    });
-
-    setHighestScore(max);
-    setMinScore(min);
-  };
-
-  // update table when turn is completed
-  const completeTurn = () => {
-    let nextTurn = turns + 1;
-    setTurns(nextTurn);
-    setCurrentTurn(nextTurn);
-    updateTable(nextTurn);
-  };
-
-  // When a table gets an update, check if game is completed, according to rules
   useEffect(() => {
     setTableFinished(handleGameState(table));
   }, [table]);
 
-  // Save Table when tableFinished is changed
   useEffect(() => {
     addTable({ ...table, tableFinished });
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [tableFinished]);
-
-  const updateTable = (rounds?: number) => {
-    const newTable: PlayGame = {
-      id: table.id,
-      rounds: rounds || turns,
-      game: table.game,
-      players: activePlayers,
-      tableFinished,
-      created: table.created,
-    };
-
-    addTable(newTable);
-  };
-
-  const addpoints = (playerID: string, turn: number, points: number | null) => {
-    if (!points && points !== 0) points = null;
-
-    const updateScoreboard = activePlayers.map((player) => {
-      if (player.id !== playerID) return player;
-
-      player.points[turn] = points;
-      player.total = player.points.reduce(
-        (sum: number, current: number | null) => {
-          if (!current) current = 0;
-          return sum + current;
-        },
-        0
-      );
-
-      player.total > highestScore
-        ? setHighestScore(player.total)
-        : updateHighScore();
-      return player;
-    });
-
-    setActivePlayers(updateScoreboard);
-  };
 
   return (
     <main className="h-full font-lucky">
@@ -148,124 +83,25 @@ export default function Play({ params }: { params: { id: string } }) {
         </section>
       ) : (
         table && (
-          <>
-            <ResizablePanelGroup direction={panelDirection}>
-              <ResizablePanel
-                defaultSize={50}
-                style={{ overflow: "auto" }}
-                minSize={panelDirection === "horizontal" ? 30 : 10}
-                className="flex flex-col"
-              >
-                <section className="flex justify-center p-2">
-                  <Link href={"/"}>
-                    <Button size={"sm"}>Back</Button>
-                  </Link>
-                  <div className="flex-1">
-                    <h2 className="mt-8 text-6xl text-center font-lucky">
-                      SKÅRBÅRD
-                    </h2>
-                    <h3 className="mx-4 text-xl text-center">
-                      {table?.game?.title}
-                    </h3>
-                  </div>
-                  <SettingsPlay />
-                </section>
-
-                <section className="flex-1 w-4/5 my-3 md:mt-12">
-                  <div className="gap-4 gap-x-0 ">
-                    {table?.players.map((player, index) => (
-                      <Bar
-                        index={index}
-                        playerName={player.name}
-                        playerScore={player.total}
-                        key={player.id}
-                        highScore={highestScore}
-                        minScore={minScore}
-                      />
-                    ))}
-                  </div>
-                </section>
-              </ResizablePanel>
-              <ResizableHandle withHandle />
-              <ResizablePanel
-                defaultSize={50}
-                minSize={panelDirection === "horizontal" ? 30 : 10}
-                className={`bg-white`}
-                style={{ overflow: "auto" }}
-              >
-                <section className="flex flex-col items-center h-full max-w-md mx-auto my-2 text-black bg-white border-none rounded-none ">
-                  <div className="grid w-full grid-cols-3 mx-auto my-8 place-items-center">
-                    {currentTurn > 0 && (
-                      <ChevronLeftIcon
-                        className="w-6 h-6 text-gray-800"
-                        onClick={() => setCurrentTurn(currentTurn - 1)}
-                      />
-                    )}
-                    <p className="col-start-2 text-3xl text-stroke shadow-slate-950 text-nowrap">
-                      Round {currentTurn + 1}
-                    </p>
-                    {currentTurn < turns && (
-                      <ChevronRightIcon
-                        className="w-6 h-6 text-gray-800"
-                        onClick={() => setCurrentTurn(currentTurn + 1)}
-                      />
-                    )}
-                  </div>
-
-                  <div className="flex flex-col items-center w-full gap-2">
-                    {currentTurn.toString() &&
-                      activePlayers.map((player) => (
-                        <div
-                          key={player.id}
-                          className="flex items-center justify-between w-full gap-2 px-4 mx-auto min-w-fit"
-                        >
-                          <p className="flex-1 w-100">{player.name}</p>
-                          <Input
-                            type="number"
-                            className="inline w-16 text-base text-center flex-0 bg-slate-100 hide-numbers"
-                            value={
-                              player.points[currentTurn] ||
-                              player.points[currentTurn] === 0
-                                ? (player.points[currentTurn] as number)
-                                : ""
-                            }
-                            onChange={(e) =>
-                              addpoints(
-                                player.id,
-                                currentTurn,
-                                parseFloat(e.target.value)
-                              )
-                            }
-                            required
-                          />
-                        </div>
-                      ))}
-
-                    <div className="flex w-full gap-4 my-4 justify-evenly">
-                      <EndGame setGameIsFinished={setTableFinished}></EndGame>
-
-                      <Button
-                        disabled={activePlayers.some((player) => {
-                          if (player.points[currentTurn] === 0) return false;
-                          if (!player.points[currentTurn]) return true;
-                        })}
-                        onClick={
-                          currentTurn === turns
-                            ? completeTurn
-                            : () => updateTable()
-                        }
-                      >
-                        {currentTurn === turns
-                          ? "Complete Turn "
-                          : "Update Turn "}
-                        {currentTurn + 1}
-                      </Button>
-                    </div>
-                  </div>
-                </section>
-              </ResizablePanel>
-            </ResizablePanelGroup>
-          </>
+          <ResizablePanelGroup direction={panelDirection}>
+            <ResizablePanel
+              defaultSize={50}
+              style={{ overflow: "auto" }}
+              minSize={panelDirection === "horizontal" ? 30 : 10}
+              className="flex flex-col"
+            >
+              <PlayLeftPanel table={table} />
+            </ResizablePanel>
+            <ResizableHandle withHandle />
+            <ResizablePanel
+              defaultSize={50}
+              style={{ overflow: "auto" }}
+              minSize={panelDirection === "horizontal" ? 30 : 10}
+              className={`bg-white`}
+            >
+              <PlayRightPanel table={table} addTable={addTable} />
+            </ResizablePanel>
+          </ResizablePanelGroup>
         )
       )}
     </main>
